requirements.txt
flask
pytesseract
pillow
pdf2image
pandas
openpyxl
langchain
faiss-cpu
PyPDF2
python-dotenv
openai
transformers 
torch
